# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: EXAMCORE PULSE API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /posts:
    get:
      operationId: getPosts
      tags: [posts]
      summary: Get all published posts
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Paginated list of posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostsPage"
    post:
      operationId: createPost
      tags: [posts]
      summary: Create a new post (admin)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePostInput"
      responses:
        "201":
          description: Post created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
...
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLoginResponse"
        "401":
          description: Invalid credentials
  /admin/posts:
    get:
      operationId: getAdminPosts
      tags: [admin]
      summary: Get all posts for admin (including drafts)
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: category
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, published, all]
            default: all
      responses:
        "200":
          description: Admin posts list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostsPage"
  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Get admin dashboard stats
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Dashboard statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    Category:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        color:
          type: string
        postCount:
          type: integer
        createdAt:
          type: string
      required:
        - _id
        - name
        - slug
...
    PostImage:
      type: object
      properties:
        url:
          type: string
        caption:
          type: string
        title:
          type: string
      required:
        - url
    Post:
      type: object
      properties:
        _id:
          type: string
        title:
          type: string
        slug:
          type: string
        excerpt:
          type: string
        content:
          type: string
        category:
          $ref: "#/components/schemas/Category"
        images:
          type: array
          items:
            $ref: "#/components/schemas/PostImage"
        coverImage:
          type: string
        status:
          type: string
          enum: [draft, published]
        views:
          type: integer
        likesCount:
          type: integer
        commentsCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - _id
        - title
        - slug
        - status
...
    PostsPage:
      type: object
      properties:
        posts:
          type: array
          items:
            $ref: "#/components/schemas/Post"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required:
        - posts
        - total
        - page
        - totalPages
    CreatePostInput:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        excerpt:
          type: string
        categoryId:
          type: string
        images:
          type: array
          items:
            $ref: "#/components/schemas/PostImage"
        coverImage:
          type: string
        status:
          type: string
          enum: [draft, published]
      required:
        - title
        - content
        - categoryId
...
    AdminLoginInput:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    AdminLoginResponse:
      type: object
      properties:
        token:
          type: string
        admin:
          type: object
          properties:
            email:
              type: string
      required:
        - token
    AdminStats:
      type: object
      properties:
        totalPosts:
          type: integer
        publishedPosts:
          type: integer
        draftPosts:
          type: integer
        totalViews:
          type: integer
        totalLikes:
          type: integer
        totalComments:
          type: integer
        totalCategories:
          type: integer
        recentPosts:
          type: array
          items:
            $ref: "#/components/schemas/Post"
      required:
        - totalPosts
        - publishedPosts
        - draftPosts
        - totalViews
        - totalLikes
        - totalComments
        - totalCategories
        - recentPosts
...