# lib/api-zod/src/generated/types/post.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EXAMCORE PULSE API
 * OpenAPI spec version: 0.1.0
 */
import type { Category } from "./category";
import type { PostImage } from "./postImage";
import type { PostStatus } from "./postStatus";

export interface Post {
  _id: string;
  title: string;
  slug: string;
  excerpt?: string;
  content?: string;
  category?: Category;
  images?: PostImage[];
  coverImage?: string;
  status: PostStatus;
  views?: number;
  likesCount?: number;
  commentsCount?: number;
  createdAt?: string;
  updatedAt?: string;
}