# lib/api-zod/src/generated/types/postImage.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EXAMCORE PULSE API
 * OpenAPI spec version: 0.1.0
 */

export interface PostImage {
  url: string;
  caption?: string;
  title?: string;
}